+++ src/components/Layout/layout.css
@@ -61,3 +61,9 @@ p {
   margin-top: 0;
   letter-spacing: 0;
 }
+
+:root {
+  --black: #000;
+  --gray-25: #bfbfbf;
+  --white: #fff
+}
