+++ src/components/Nav/Nav.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { ArrowLeftIcon } from '../Icons'
 import { Link } from 'gatsby'
 import { StyledNav } from './styles'
-import { useBreakpoint } from '../../hooks/useBreakpoints'
+import { useBreakpoint } from '../../lib/hooks/useBreakpoints'
 
 export default function Nav({data, title}) {
   const [isActive, setIsActive] = useState(false)
