+++ src/components/Project/Project.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import {StyledProject, StyledProjectInfoContainer, StyledPhotosContainer, DivPhoto} from './styles'
 import Lightbox from '../Lightbox/Lightbox'
-import { useBreakpoint } from "../../hooks/useBreakpoints"
+import { useBreakpoint } from "../../lib/hooks/useBreakpoints"
 import Nav from '../Nav/Nav'
 
 export default function Project({project}) {
