+++ src/lib/hooks/usePanAndZoom.js
